server/requirements.txt
Flask==3.0.0
flask_cors==4.0.0
nltk==3.8.1
pdfminer.six==20221105
scikit_learn==1.3.2
